third_party/WebKit/LayoutTests/editing/deleting/skip-virama-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable>|\u0915\u094D\u0937</div>',
        selection => {
            assert_not_equals(window.eventSender, undefined,
                'This test requires eventSender');
            eventSender.keyDown('Backspace', null);
        },
        '<div contenteditable>|\u0915\u094D\u0937</div>'),
        'prevent a cursor from moving after a virama sign');
</script>
